29-add-events-integration: refactored amqp eventing packages
