+++ cm/cm-biz-local/src/main/java/com/dj/cm/biz/service/echo/impl/EchoBizEventListener.java
@@ -1,7 +1,7 @@
 package com.dj.cm.biz.service.echo.impl;
 
-import com.dj.cm.event.config.EventConfig;
-import com.dj.cm.event.echo.EchoEvent;
+import com.dj.cm.event.amqp.config.AmqpEventConfig;
+import com.dj.cm.event.model.echo.EchoEvent;
 import org.slf4j.Logger;
 import org.springframework.amqp.rabbit.annotation.RabbitListener;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,7 +15,7 @@ public class EchoBizEventListener {
 	@Autowired
 	private Logger logger;
 
-	@RabbitListener(queues = EventConfig.ECHO_BIZ_QUEUE_NAME)
+	@RabbitListener(queues = AmqpEventConfig.ECHO_BIZ_QUEUE_NAME)
 	public void handle(EchoEvent event) {
 		logger.info(" [x] Received '" + event + "'");
 	}
