--- cm/cm-event/src/main/java/com/dj/cm/event/echo/EchoEvent.java
@@ -1,11 +1,0 @@
-package com.dj.cm.event.echo;
-
-
-import com.dj.cm.event.common.Event;
-
-public class EchoEvent extends Event {
-
-	public EchoEvent(EventType type) {
-		super(type);
-	}
-}
