+++ cm/cm-event/src/main/java/com/dj/cm/event/model/common/EventType.java
@@ -1,0 +1,5 @@
+package com.dj.cm.event.model.common;
+
+public enum EventType {
+	CREATED, UPDATED, DELETED;
+}
