+++ cm/cm-event/src/main/java/com/dj/cm/event/model/echo/EchoCreatedEvent.java
@@ -1,6 +1,7 @@
-package com.dj.cm.event.echo;
+package com.dj.cm.event.model.echo;
 
 
+import com.dj.cm.event.model.common.EventType;
 import com.dj.cm.model.entity.Echo;
 
 public class EchoCreatedEvent extends EchoEvent {
