+++ cm/cm-event/src/main/java/com/dj/cm/event/model/echo/EchoEvent.java
@@ -1,0 +1,12 @@
+package com.dj.cm.event.model.echo;
+
+
+import com.dj.cm.event.model.common.Event;
+import com.dj.cm.event.model.common.EventType;
+
+public class EchoEvent extends Event {
+
+	public EchoEvent(EventType type) {
+		super(type);
+	}
+}
